Shahid Arabic Update

+++ plugin.video.shahidmbcnet/default.py
@@ -9,7 +9,7 @@ import os
 import sys
 from BeautifulSoup import BeautifulStoneSoup, BeautifulSoup, BeautifulSOAP, Tag,NavigableString
 try:
-  from lxml import etree
+  from lxmlERRRORRRR import etree
   print("running with lxml.etree")
 except ImportError:
 	try:
@@ -38,7 +38,8 @@ addon_id = 'plugin.video.shahidmbcnet'
 selfAddon = xbmcaddon.Addon(id=addon_id)
 addonPath = xbmcaddon.Addon().getAddonInfo("path")
 addonArt = os.path.join(addonPath,'resources/images')
-communityStreamPath = os.path.join(addonPath,'resources/community')
+communityStreamPath = os.path.join(addonPath,'resources')
+communityStreamPath =os.path.join(communityStreamPath,'community')
 profile_path =  xbmc.translatePath(selfAddon.getAddonInfo('profile'))
  
 mainurl='http://shahid.mbc.net'
@@ -77,9 +78,13 @@ VIEW_MODES = {
 }
 
 def get_view_mode_id( view_mode):
-	view_mode_ids = VIEW_MODES.get(view_mode.lower())
-	if view_mode_ids:
-		return view_mode_ids.get(xbmc.getSkinDir())
+	default_view_mode=selfAddon.getSetting( "usethisviewmode" )
+	if default_view_mode=="":
+		view_mode_ids = VIEW_MODES.get(view_mode.lower())
+		if view_mode_ids:
+			return view_mode_ids.get(xbmc.getSkinDir())
+	else:
+		return int(default_view_mode)
 	return None
 
 def addLink(name,url,iconimage):
@@ -103,7 +108,7 @@ def Colored(text = '', colorid = '', isBold = False):
 		text = '[B]' + text + '[/B]'
 	return '[COLOR ' + color + ']' + text + '[/COLOR]'	
 	
-def addDir(name,url,mode,iconimage	,showContext=False,isItFolder=True,pageNumber="", isHTML=True,addIconForPlaylist=False, AddRemoveMyChannels=None):
+def addDir(name,url,mode,iconimage	,showContext=False,isItFolder=True,pageNumber="", isHTML=True,addIconForPlaylist=False, AddRemoveMyChannels=None, SelectDefaultSource=None, hideChannel=None):
 #	print name
 #	name=name.decode('utf-8','replace')
 	if isHTML:
@@ -136,13 +141,36 @@ def addDir(name,url,mode,iconimage	,showContext=False,isItFolder=True,pageNumber
 		cmd3 = "XBMC.RunPlugin(%s&cdnType=%s)" % (u, "ak")
 		liz.addContextMenuItems([('Play using L3 Cdn',cmd1),('Play using XDN Cdn',cmd2),('Play using AK Cdn',cmd3)])
 
+	context_menu=[]
 	if not AddRemoveMyChannels==None:
 		if AddRemoveMyChannels:
 			cmd1 = "XBMC.RunPlugin(%s&AddRemoveMyChannels=add)" % (u)
-			liz.addContextMenuItems([('Add to My Channels',cmd1)])
+			context_menu.append(('Add to My Channels',cmd1))
 		else:
 			cmd1 = "XBMC.RunPlugin(%s&AddRemoveMyChannels=remove)" % (u)
-			liz.addContextMenuItems([('Remove from My Channels',cmd1)])
+			context_menu.append(('Remove from My Channels',cmd1))
+
+
+	if SelectDefaultSource:
+		#print 'select defauly'
+		cmd2 = "XBMC.RunPlugin(%s&selectDefaultSource=yes)" % (u)
+		context_menu.append(('Select default source',cmd2))
+
+	if not hideChannel==None:
+		if hideChannel:
+			cmd3 = "XBMC.RunPlugin(%s&HideChannel=yes)" % (u)
+			context_menu.append(('Hide this Channel',cmd3))
+		else:
+			cmd3 = "XBMC.RunPlugin(%s&HideChannel=no)" % (u)
+			context_menu.append(('Unhide this Channel',cmd3))
+
+            
+	if len(context_menu)>0:
+		liz.addContextMenuItems(context_menu,replaceItems=False)
+	#	if selfAddon.getSetting( "addToFavMode" )=="true":
+	#		liz.addContextMenuItems(context_menu,replaceItems=False)
+	#	else:
+	#		liz.addContextMenuItems(context_menu,replaceItems=True)
 			
 	ok=xbmcplugin.addDirectoryItem(handle=int(sys.argv[1]),url=u,listitem=liz,isFolder=isItFolder)
 	return ok
@@ -324,21 +352,47 @@ def RefreshResources(auto=False):
 		fileno+=1
 	pDialog.close()
 	dialog = xbmcgui.Dialog()
-	ok = dialog.ok('XBMC', 'Download finished. Close close Addon and come back')
+	ok = dialog.ok('XBMC', 'Download finished. Close Addon and come back')
 
-def removeLoginFile():
+def removeLoginFile(livePlayer,TeleDunet):
+	something_done=False
 	try:
-		COOKIEFILE = communityStreamPath+'/livePlayerLoginCookie.lwp'
-		os.remove(COOKIEFILE)
+		if livePlayer:
+			something_done=True
+			selfAddon.setSetting( id="lastLivetvWorkingCode" ,value="")
+			COOKIEFILE = communityStreamPath+'/livePlayerLoginCookie.lwp'
+			os.remove(COOKIEFILE)
+			
 	except: pass
 	try:
-		COOKIEFILE = communityStreamPath+'/teletdunetPlayerLoginCookie.lwp'
-		os.remove(COOKIEFILE)
+		if TeleDunet:
+			if communityStreamPath not in sys.path:
+				sys.path.append(communityStreamPath)
+			#print processor
+		
+		
+			#from importlib import import_module
+			processorObject=import_module('teledunetPlayer')
+			processorObject.clearFileCache()
+			something_done=True
+			COOKIEFILE = communityStreamPath+'/teletdunetPlayerLoginCookie.lwp'
+			os.remove(COOKIEFILE)
 	except: pass
+	if something_done:
+		time = 2000  #in miliseconds
+		line1="Session data removed!"
+		xbmc.executebuiltin('Notification(%s, %s, %d, %s)'%(__addonname__,line1, time, __icon__))
 
 def ShowSettings(Fromurl):
+	current_LivePlayerLogin=selfAddon.getSetting( "liveTvLogin" )+selfAddon.getSetting( "liveTvPassword")
+	current_teleDunetLogin=selfAddon.getSetting( "teledunetTvLogin" )+selfAddon.getSetting( "teledunetTvPassword")
+	selfAddon.setSetting( id="clearLogonSettings" ,value="false")
 	selfAddon.openSettings()
-	removeLoginFile()
+	#print 'after settings'
+	clearLogonSettings=selfAddon.getSetting( "clearLogonSettings" )
+	after_LivePlayerLogin=selfAddon.getSetting( "liveTvLogin" )+selfAddon.getSetting( "liveTvPassword")
+	after_teleDunetLogin=selfAddon.getSetting( "teledunetTvLogin" )+selfAddon.getSetting( "teledunetTvPassword")
+	removeLoginFile(clearLogonSettings=="true" or not current_LivePlayerLogin==after_LivePlayerLogin, clearLogonSettings=="true" or not current_teleDunetLogin==after_teleDunetLogin )
 	return
 	
 def AddSeries(Fromurl,pageNumber=""):
@@ -592,7 +646,7 @@ def getFirstElement(elements,attrib, val):
 	for el in elements:
 		#print el.attrib[attrib]
 		if el.attrib[attrib]==val:
-			print 'found next'
+			#print 'found next'
 			return el
 	return None
 
@@ -745,14 +799,36 @@ def getSourceAndStreamInfo(channelId, returnOnFirst,pDialog):
 		ret=[]
 		#Ssoup=getSoup('Sources.xml');
 		sourcesXml=getEtreeFromFile('Sources.xml');
+		default_source=''
+		config=getChannelSettings( channelId)
+		match_title=''
+		if config and 'defaultsource' in config:
+			default_source=config['defaultsource'].split(':')[0]
+			try:
+				match_title= ''.join(config['defaultsource'].split(':')[1:])
+				print 'match_title in settings',match_title,default_source
+			except: pass
+		#print 'default_source',default_source
 		orderlist={}
-		for n in range(6):
+		default_source_exists=False
+		total_sources=12
+		for n in range(total_sources):
 			val=selfAddon.getSetting( "order"+str(n+1) )
 			if val and not val=="":
+				#print 'val',val,default_source
 				orderlist[val]=n*100
+				if not default_source=='' and default_source ==val:
+					orderlist[val]=-100
+
+
+
+					
+			
 		#print orderlist
 		#print 'sources',sources
 		num=0
+
+		
 		pDialog.update(30, 'Looping on sources')
 		sources=sourcesXml.findall('source')
 		for source in sources:
@@ -791,17 +867,41 @@ def getSourceAndStreamInfo(channelId, returnOnFirst,pDialog):
 					sInfo=[]
 					for inf in sInfos:
 						if inf.findtext('cname').lower()==channelId.lower():
+							source_title=''
+							if match_title<>'':
+								try:
+									if source.findtext('sname')=='generic':
+										source_title=inf.find('item').findtext('title')
+									else:
+										source_title=inf.findtext('title')
+								except: pass
+                                    
+							#print default_source,sid,match_title,inf.findtext('title'),inf
+							if not default_source=='' and default_source==sname and (match_title =='' or match_title==source_title):                       
+								default_source_exists=True
 							sInfo.append(inf)
 					name_find=sname
 					if name_find in orderlist:
 						order= orderlist[name_find]
 					else:
-						order=1000
+						print 'not found',name_find,orderlist
+						order=20000
 					order+=num
 					if not sInfo==None and len(sInfo)>0:
-						#print 'sInfo...................',len(sInfo)
+						print 'sInfo...................',len(sInfo)
 						
 						for single in sInfo:
+							source_title=''
+							if match_title<>'':
+								try:
+									if source.findtext('sname')=='generic':
+										source_title=single.find('item').findtext('title')
+									else:
+										source_title=single.findtext('title')
+								except: pass
+							if (match_title =='' or match_title==source_title):
+								print 'title match for order', match_title,source_title
+								order-=1
 							ret.append([source,single,order])
 						#if returnOnFirst:
 						#	break;
@@ -812,9 +912,13 @@ def getSourceAndStreamInfo(channelId, returnOnFirst,pDialog):
 		traceback.print_exc(file=sys.stdout)
 		pass
 	#print 'unsorted ret',ret
-	return sorted(ret,key=lambda x:x[2])
 
-def selectSource(sources):
+	#print ret
+	ret= sorted(ret,key=lambda x:x[2])
+	print ret
+	return ret,default_source_exists and not default_source==''
+
+def selectSource(sources,fromSelectSource=False):
     if 1==1 or len(sources) > 1:
         #print 'total sources',len(sources)
         dialog = xbmcgui.Dialog()
@@ -824,19 +928,60 @@ def selectSource(sources):
             #print 'i',i.id,i
             if s.findtext('id')=="generic":
                 try:
-                    print 'trying generic name'
+                    #print 'trying generic name'
                     titles.append(s.findtext('sname')+': '+i.find('item').findtext('title'))
-                    print 'trying generic name end '
+                    #print 'trying generic name end '
                 except:
                     titles.append(s.findtext('sname'))
             else:
-                titles.append(s.findtext('sname'))
+                try:
+                    titles.append(s.findtext('sname')+': '+i.findtext('title'))
+                except: titles.append(s.findtext('sname'))
+
+        if fromSelectSource:
+            titles.append(Colored('Clear Default Source setting','one',True))
+
         index = dialog.select('Choose your stream', titles)
         if index > -1:
-            return sources[index]
+            if index>len(sources)-1:
+                return 'remove'#remove it
+            else:
+                return sources[index]
         else:
             return False
 
+def selectDefaultSourcesForChannel(channelId ):
+	try:
+		pDialog = xbmcgui.DialogProgress()
+		ret = pDialog.create('XBMC', 'Finding available resources...')
+		pDialog.update(20, 'Finding sources..')
+		providers, default_source_exists=getSourceAndStreamInfo(channelId,False,pDialog)
+		if len(providers)==0:
+			pDialog.close()
+			time = 2000  #in miliseconds
+			line1="No sources found"
+			xbmc.executebuiltin('Notification(%s, %s, %d, %s)'%(__addonname__,line1, time, __icon__))
+			return None
+		selectedprovider=selectSource(providers,True)
+		if not selectedprovider:
+			return None
+		if selectedprovider=='remove':
+			return ''
+		fav_source=''   
+		source,sInfo,order=selectedprovider #pick first one
+		fav_source=source.findtext('sname')
+		try:
+			if source.findtext('id')=="generic":
+				fav_source=source.findtext('sname')+':'+sInfo.find('item').findtext('title')
+			else:
+				fav_source=source.findtext('sname')+':'+sInfo.findtext('title')
+		except:pass
+
+		return fav_source
+	except:
+		traceback.print_exc(file=sys.stdout)
+		return None
+		
 def PlayCommunityStream(channelId, name, mode):
 	try:
 		#print 'PlayCommunityStream'
@@ -851,7 +996,9 @@ def PlayCommunityStream(channelId, name, mode):
 			playFirst=True
 		playFirst=bool(playFirst)
 		pDialog.update(20, 'Finding sources..')
-		providers=getSourceAndStreamInfo(channelId,playFirst,pDialog)
+		providers,default_source_exists=getSourceAndStreamInfo(channelId,playFirst,pDialog)
+		if default_source_exists:
+			playFirst=True
 		if len(providers)==0:
 			pDialog.close()
 			time = 2000  #in miliseconds
@@ -1008,7 +1155,7 @@ def addToMyChannels(cname):
 		MyChannelList= BeautifulSOAP('<channels></channels>')
 	
 	val=MyChannelList.find("channel",{"cname":cname})
-	print 'val is ',val
+	#print 'val is ',val
 	if not val:
 		channeltag = Tag(MyChannelList, "channel")
 		channeltag['cname']=cname
@@ -1017,7 +1164,7 @@ def addToMyChannels(cname):
 		#cnametag.insert(0, ctext)
 		#channeltag.insert(0, cnametag)
 		MyChannelList.channels.insert(0, channeltag)
-		print MyChannelList.prettify()
+		#print MyChannelList.prettify()
 
 		with open(fileName, "wb") as filewriter:
 			filewriter.write(str(MyChannelList))
@@ -1033,7 +1180,7 @@ def removeFromMyChannels(cname):
 	
 	val=MyChannelList.find("channel",{"cname":cname})
 	if val:
-		print 'val to be deleted',val
+		#print 'val to be deleted',val
 		val.extract()
 
 		with open(fileName, "wb") as filewriter:
@@ -1042,7 +1189,7 @@ def removeFromMyChannels(cname):
 def addCommunityCats():
 	#soup=getSoup('Categories.xml');
 	cats=getEtreeFromFile('Categories.xml');
-#	print cats 
+	#print cats 
 
 	addDir('My Channels' ,'My Channels' ,15,addonArt+'/mychannels.png', False,isItFolder=True)		#name,url,mode,icon
 
@@ -1063,20 +1210,30 @@ def getCommunityChannels(catType):
 	searchCall='channel'
 	#if not catType=="all":
 	searchCall='.//category'
-	print searchCall
+	#print searchCall
 	MyChannelList=None
+	hidechanneloption=True
+	sourcesXml=getEtreeFromFile('Sources.xml');
+	sources_list={}
+	for sources in sourcesXml.findall('source'): 
+		sname=sources.findtext('sname')
+		ssname=sources.findtext('shortname')
+		scolour=sources.findtext('colour')
+		sources_list[sname]=[ssname,scolour]
 	if catType=="My Channels":
 		try:
 			fileName=os.path.join(profile_path, 'MyChannels.xml')
-			print fileName
+			#print fileName
 			MyChannelList=getSoup(fileName,True)
-			print MyChannelList
+			#print MyChannelList
 		except: MyChannelList=None
 		
 	for channel in Channelsxml.findall('channel'):
 		#print channel
 		chName=channel.findtext('cname')
 		if 1==1:
+			config=getChannelSettings( chName)
+			#print 'config is ',config
 			if not catType=="all":
 				exists=False
 				if not catType=="My Channels":
@@ -1094,14 +1251,31 @@ def getCommunityChannels(catType):
 						val=MyChannelList.find("channel",{"cname":chName})
 						if val:
 							exists=True
+				if config and 'hidden' in config:
+					exists=not config['hidden']=="yes"
 				if not exists:
 					continue
+
 			
 
 		
+		if config and 'hidden' in config:
+			hidechanneloption=not config['hidden']=="yes"
 		#chUrl = channel.id.text
 		imageUrl =channel.findtext('imageurl')
- 		retVal.append([chName,chName,imageUrl])
+		chUrl=chName
+		if config and 'defaultsource' in config:
+			default_source=config['defaultsource'].split(':')[0]
+			#chName+='['+default_source+']'
+			if default_source in sources_list:
+				short_name='['+sources_list[default_source][0]+']'
+				colour=sources_list[default_source][1]
+				short_name=Colored(text = short_name, colorid = colour, isBold = False)
+				print short_name
+
+				chName+=' '+short_name
+				
+ 		retVal.append([chUrl,chName,imageUrl,hidechanneloption])
 	return retVal
 	
 
@@ -1112,17 +1286,54 @@ def addCommunityChannels(catType):
 		chName=channel[1]
 		chUrl = channel[0]
 		imageUrl = channel[2]
+		hideChannel=channel[3]
 		addRemoveMyChannel=not catType=="My Channels"
- 		addDir(chName ,chUrl ,16,imageUrl, False,isItFolder=False,AddRemoveMyChannels=addRemoveMyChannel)		#name,url,mode,icon
+ 		addDir(chName ,chUrl ,16,imageUrl, False,isItFolder=False,AddRemoveMyChannels=addRemoveMyChannel, SelectDefaultSource=True,hideChannel=hideChannel )		#name,url,mode,icon
 	return
 
-def getEtreeFromFile(fileName, isabsolutePath=False):
-	strpath=os.path.join(communityStreamPath, fileName)
-	if isabsolutePath:
-		strpath=fileName
-	data = open(strpath, "r").read()
-	return getETreeFromString(data)
+def setChannelSettings(cname,settingName,SettingVal):
+	current_setting=getChannelSettings(cname)
+	if current_setting==None:
+		current_setting={settingName:SettingVal}
+	else:
+		current_setting[settingName]=SettingVal
+	#print 'current_setting',current_setting
+	saveChannelSettings(cname,current_setting)
 
+
+def getChannelSettings(cname):
+	current_string=selfAddon.getSetting( cname+"-settings")
+	#print cname+"-settings",current_string
+	if current_string=="":
+		return None
+	#print 'current_string',current_string
+	return json.loads(current_string)
+
+
+def saveChannelSettings(cname, json_data):
+	store=""
+	if json_data:
+		store=json_data
+	selfAddon.setSetting( id=cname+"-settings" ,value=json.dumps(store))
+
+
+
+	
+	
+def getEtreeFromFile(fileName, isabsolutePath=False):
+	try:
+		#print 'communityStreamPath',communityStreamPath
+		#print 'fileName',fileName
+		strpath=os.path.join(communityStreamPath, fileName)
+		#print 'strpath',strpath
+		if isabsolutePath:
+			strpath=fileName
+		data = open(strpath, "r").read()
+		return getETreeFromString(data)
+	except:
+		print 'somethingwrong'
+		traceback.print_exc(file=sys.stdout)
+	
 #obselete
 def getSoup(fileName, isabsolutePath=False):
 	strpath=os.path.join(communityStreamPath, fileName)
@@ -1404,9 +1615,20 @@ try:
 except:
 	pass
 
+selectDefaultSource=None
+
+try:
+	selectDefaultSource=args.get('selectDefaultSource', None)[0]
+except:
+	pass
+
+HidChannel=None
+try:
+	HidChannel=args.get('HideChannel', None)[0]
+except:
+	pass
+
 	
-
-
 
 print 	mode,pageNumber
 
@@ -1426,6 +1648,42 @@ try:
 			mode=15
 			url="My Channels"
 			print mode
+
+	if not HidChannel==None:
+		if HidChannel=="yes":
+			setChannelSettings(url,'hidden',HidChannel)
+			line1 = 'Channel has been hidden in categories'
+			time = 2000  #in miliseconds
+			xbmc.executebuiltin('Notification(%s, %s, %d, %s)'%(__addonname__,line1, time, __icon__))
+			mode=-1
+		else:
+			setChannelSettings(url,'hidden','')
+			line1 = 'Channel has been unhidden in categories'
+			time = 2000  #in miliseconds
+			xbmc.executebuiltin('Notification(%s, %s, %d, %s)'%(__addonname__,line1, time, __icon__))
+			mode=-1
+			url=""
+			
+	if not selectDefaultSource==None:
+		default_source=selectDefaultSourcesForChannel(url)
+		print 'v',default_source
+		if not default_source==None:
+			print 'saving settings',default_source    
+			setChannelSettings(url,'defaultsource',default_source)
+			line1 = 'setting saved'
+			time = 2000  #in miliseconds
+			xbmc.executebuiltin('Notification(%s, %s, %d, %s)'%(__addonname__,line1, time, __icon__))
+		mode=-1
+
+	if mode==299: #add communutycats
+		print 'delete cache'
+		removeLoginFile(True,True)
+		line1 = 'Login sessions cleared!'
+		time = 2000  #in miliseconds
+		xbmc.executebuiltin('Notification(%s, %s, %d, %s)'%(__addonname__,line1, time, __icon__))
+		xbmc.executebuiltin('Container.Update(plugin://plugin.video.shahidmbcnet)')
+		mode=-1
+		
 	if mode==None or url==None or len(url)<1:
 		print "InAddTypes"
 		checkAndRefresh()        
@@ -1484,6 +1742,7 @@ try:
 	elif mode==23: #add communutycats
 		print "play youtube url is "+url,mode
 		AddYoutubeVideosByPlaylist(url);	
+
 except:
 	print 'somethingwrong'
 	traceback.print_exc(file=sys.stdout)
